refactor(web): extract sub-tasks-section from task detail
